UPDATES: adding compatibility to EIP2981, fixing typos and formating code

+++ contracts/Royalties-registry/IRoyaltiesProvider.sol
@@ -44,5 +44,7 @@ pragma solidity ^0.8.0;
 import "../LibPart.sol";
 
 interface IRoyaltiesProvider {
-    function getRoyalties(address token, uint tokenId) external returns (LibPart.Part[] memory);
+    function getRoyalties(address token, uint256 tokenId)
+        external
+        returns (LibPart.Part[] memory);
 }
