+++ contracts/exchange/Luxy.sol
@@ -50,14 +50,18 @@ contract Luxy is LuxyCore, LuxyTransferManager {
     function __LuxyCore_init(
         INftTransferProxy _transferProxy,
         IERC20TransferProxy _erc20TransferProxy,
-        uint newProtocolFee,
+        uint256 newProtocolFee,
         address newDefaultFeeReceiver,
         IRoyaltiesProvider newRoyaltiesProvider
     ) external initializer {
         __Context_init_unchained();
         __Ownable_init_unchained();
         __TransferExecutor_init_unchained(_transferProxy, _erc20TransferProxy);
-        __LuxyTransferManager_init_unchained(newProtocolFee, newDefaultFeeReceiver, newRoyaltiesProvider);
-        __OrderValidator_init_unchained('LuxyValidator','1');
+        __LuxyTransferManager_init_unchained(
+            newProtocolFee,
+            newDefaultFeeReceiver,
+            newRoyaltiesProvider
+        );
+        __OrderValidator_init_unchained("LuxyValidator", "1");
     }
 }
