+++ contracts/exchange/lib/LibMath.sol
@@ -44,7 +44,7 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts-upgradeable/utils/math/SafeMathUpgradeable.sol";
 
 library LibMath {
-    using SafeMathUpgradeable for uint;
+    using SafeMathUpgradeable for uint256;
 
     /// @dev Calculates partial value given a numerator and denominator rounded down.
     ///      Reverts if rounding error is >= 0.1%
@@ -103,11 +103,7 @@ library LibMath {
         //        1000 * remainder  <  numerator * target
         // so we have a rounding error iff:
         //        1000 * remainder  >=  numerator * target
-        uint256 remainder = mulmod(
-            target,
-            numerator,
-            denominator
-        );
+        uint256 remainder = mulmod(target, numerator, denominator);
         isError = remainder.mul(1000) >= numerator.mul(target);
     }
 
@@ -119,7 +115,9 @@ library LibMath {
         if (isRoundingErrorCeil(numerator, denominator, target)) {
             revert("rounding error");
         }
-        partialAmount = numerator.mul(target).add(denominator.sub(1)).div(denominator);
+        partialAmount = numerator.mul(target).add(denominator.sub(1)).div(
+            denominator
+        );
     }
 
     /// @dev Checks if rounding error >= 0.1% when rounding up.
@@ -144,11 +142,7 @@ library LibMath {
             return false;
         }
         // Compute remainder as before
-        uint256 remainder = mulmod(
-            target,
-            numerator,
-            denominator
-        );
+        uint256 remainder = mulmod(target, numerator, denominator);
         remainder = denominator.sub(remainder) % denominator;
         isError = remainder.mul(1000) >= numerator.mul(target);
         return isError;
