+++ contracts/tokens/testContracts/TestERC1155.sol
@@ -4,14 +4,17 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts-upgradeable/token/ERC1155/ERC1155Upgradeable.sol";
 
 contract TestERC1155 is ERC1155Upgradeable {
-
     function __TestERC1155_init(string memory uri_) public initializer {
         __Context_init_unchained();
         __ERC165_init_unchained();
         __ERC1155_init_unchained(uri_);
     }
-    function mint(address to, uint tokenId, uint amount) external {
+
+    function mint(
+        address to,
+        uint256 tokenId,
+        uint256 amount
+    ) external {
         _mint(to, tokenId, amount, "");
     }
-
 }
