+++ contracts/tokens/testContracts/TestERC20.sol
@@ -4,11 +4,15 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol";
 
 contract TestERC20 is ERC20Upgradeable {
-    function __TestERC20_init(string memory name_, string memory symbol_)  public initializer {
+    function __TestERC20_init(string memory name_, string memory symbol_)
+        public
+        initializer
+    {
         __Context_init_unchained();
         __ERC20_init_unchained(name_, symbol_);
     }
-    function mint(address to, uint amount) external {
+
+    function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
 }
