+++ contracts/tokens/testContracts/TestERC721.sol
@@ -4,12 +4,16 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts-upgradeable/token/ERC721/ERC721Upgradeable.sol";
 
 contract TestERC721 is ERC721Upgradeable {
-    function __TestERC721_init(string memory name_, string memory symbol_) public initializer {
+    function __TestERC721_init(string memory name_, string memory symbol_)
+        public
+        initializer
+    {
         __Context_init_unchained();
         __ERC165_init_unchained();
         __ERC721_init_unchained(name_, symbol_);
     }
-    function mint(address to, uint tokenId) external {
+
+    function mint(address to, uint256 tokenId) external {
         _mint(to, tokenId);
     }
 }
